requirements.txt
Alfred-Workflow
scrapy
